implemented manual controls, renamed alignLiquidity()
I received a lot of feedback about how the alignLiquidity() function was too restrictive and didn't allow for partial liquidity additions. The all-in-one function could create a lot of MEV, so doing it in batches could be desirable. I implemented alignNfts() and alignTokens() to help handle this. alignNfts() takes a uint256 array of tokenIds that the vault should be in possession of and attempts to add them to NFTX liquidity if enough ETH is held. alignTokens() takes a uint256 input for how much WETH in wei to add to NFTX liquidity alongside all NFTX inventory tokens. additionally, I implemented getInventory() which will return the entire nftsHeld array, allowing the contract owner to easily see what NFTs the contract is aware of. This will help inform if they need to run checkInventory().